[fshost] add more logging to rust fshost
Log when calls to the fshost admin and block watcher protocols are made
or processed. For admin, the logging is done when the call is made, for
block watcher, they are printed when the pause or resume is completed.

This also modifies logs related to block device processing to print out
the class path instead of the full topological path for log messages
following the initial matching message. It's long and tricky to read
quickly.

Bug: 87511
Bug: 112161
Change-Id: I675e14ff6a72d797706f2659223445748febef86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767469
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>