[fvm] Unpack command verifies image length to warn of mismatch.
Because the underlying reads are preads it will silently return zeroes
when reading past the end of the image. Dealing with the resulting
images can be confusing. Adding a warning message when the image size is
not as expected.

Tested manually with a known bad and a known good image from production.

Bug: b/242716290
Change-Id: I9a548674624449c3ec056ef5fcb6e8eeac3fd724
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713682
Reviewed-by: Mike McCreavy <mccreavy@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>