[lib][fake-object] Use _zx symbols for real syscalls
This is essentially a manual revert of:
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/362573

Per discussion with Roland, using dlsym remains fragile and we should be
able to rely on the _zx* symbols now. This resolves an issue where if an
object initialized in global cxx scope made a syscall the calls to
dlopen/dlsym would hang.

Change-Id: Ifb6b09dfb9eced30917c380f1ac65a3468a67aa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768305
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Christopher Anderson <cja@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>