[Settings] Remove empty() for DND
The original TODO was to update to default() so it is consistent with Rust convention. But discussed offline during stand up and found the usage of this function can be replaced with a direct construction. Therefore, update the code and delete the TODO.

Fixed: 67780
Test: fx test setui_service_tests
Change-Id: Ie682d556f280f9134e51159c441141785a0e4cfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766679
Fuchsia-Auto-Submit: Sijie Chen <sijiec@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>