[netemul] Log errors from devfs instead of returning
When netemul is creating a new ethernet endpoint, it watches
/dev/class/ethernet to wait for an ethertap device to come up that has
the MAC address of the device we created. In tests with multiple
ethertap endpoints, one endpoint might be in the process of shutting
down while netemul is waiting for another one to come up. This means
that while it's watching devfs, it might try and fail to interact with
that device that is shutting down.

Because it's expected that we will see potentially nonresponsive devices
mid-teardown, rather than returning an error in this situation, we
should log a warning and keep watching the directory for the device
we're looking for.

Fixed: 112823, 116538, 116543, 116537
Multiply: netstack-interfaces-fidl-test
Change-Id: Ied8278e072440b50e2edaa2eb3b40298467010b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770203
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>