[modular] Reconnect to Agent protocol served by v2_modular_agents
Previously, sessionmgr would connect to the fuchsia.modular.Agent
protocol exposed by a v2_modular_agent once, on startup. If that
connection was closed, sessionmgr would incorrectly try to restart
that agent as a v1 component.

sessionmgr now reconnects to Agent if the connection is closed,
usually as a result of the v2 agent component terminating.

Fixed: 116254
Change-Id: I3be69cf2d47b3d31cf1db4b20702a9e669f09355
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766677
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>