[ffx][component] Fix ffx component copy wildcard bug for nested directories
This CL fixes the issue with wildcard support for when a namespace
directory contains both files and directories. Originally, the paths of
directories were being passed as valid source paths to component copy
during remote wildcard expansion. This CL prints a message to inform the
user that the path will be ignored.

Change-Id: I0eca68a647641d3b7a29373b36eba7568f371812
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765727
Commit-Queue: Richard Chen <richardrchen@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>