[kms-stateless] Ignore empty filename in WatchTee
If the file watcher is activated before the optee driver is ready, it
will can return an empty filename as a result for the watcher to
process. If this isn't guarded against, WatchTee will proceed to make a
request to the optee driver, which will fail and trigger an immediate
retry.

The result is a race condition in which fshost asks kms-stateless for
the device key to unseal zxcrypt for minfs before optee is ready. After
5 retries, fshost gives up and leaves the device in a state where minfs
  remains encrypted.

Adding this check allows the watcher to continue watching for the actual
device to show up instead of trying to proceed with an empty value.

Bug: b/260904369
Test: Manual
Change-Id: I987f4f1e60ea0410a4bb27f93cbf9313bb5d689e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770386
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>