[bundles] Remove non-test targets from '//src/**:tests' groups
This is a net-0 change to the set of tests included by
//bundles/buildbot/core. [1]

This changes //bundles/buildbot/core to include //src:tests_no_e2e
and //src:e2e_tests instead of the //src group.  These groups
only dep on on the various areas ':tests' groups, which should only
directly dep on tests and sub-area '<foo>:tests' groups.
Production software should not be in the direct dependencies of any
of these groups (but the tests in question _should_ dep on the
product components, to make sure that changes to production code
marks tests as needing to be re-run).

This ensures that including '//bundles/buildbot/core' does not
include subsystems that the product itself does not include.

[1] This was verified by comparing the list of GN labels for
tests in 'tests.json' before and after the change was made.

Bug: 114931
Change-Id: I4ccc38e8714e450e02792036c5a90f25781c53c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758059
Reviewed-by: Zach Bush <zmbush@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>