[starnix] Allow memory reclamation in binder driver
This CL replaces the previous allocator that did not correctly reclaim
memory for a new one that is inneficient (allocation are in O(n) in the
number of previous allocations) but reclaim memory.

This should be updated to a better allocator, but should help prevent
ENOMEM issues for now.

Change-Id: Iab37778d75a43447847a7ad6a78dcc8f790f43a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770862
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>