[ethernet] Deflake test
Because multiple ethertap devices are created and torn down concurrently
in this test, it's possible for the test to observe a PEER_CLOSED error
interacting with a device that is mid-teardown. In this scenario, keep
watching for devices rather than propagating the error.

Fixed: 116578
Multiply: ethernet-integration-test
Change-Id: I7622e9ed5e45bc4a4061887e01abf97f58b7ecb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771003
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>