[tpm] Add //src/security/lib/tpm.
This new tpm library is the Fuchsia specific interface point for
consumers of the low level TPM interface. This is an abstraction
ontop of the TCG TSS2 interface that exposes just the
functionality we need to use in the Fuchsia project through
safe Rust interfaces. The underlying TCG TSS2 interface is written
in C and we export a Rust interface via bindgen. However this
interface still is unsafe and so this high level wrapper takes
the large exported tpm2-tss-sys interface and exposes a much
smaller subset in a Rust idomatic way that prevents accidental
misuse.

Change-Id: I163d6c8a8c5b11ec1b613339926879818cf4ccbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759263
Reviewed-by: Aaron Green <aarongreen@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Ben Wright <benwright@google.com>