[modular] abort initialization if sessionmgr was asked to terminate.
This is an attempt to fix fxb/115009, which resulted in a use-after-free from a call to `GetAgentOutgoingServices()` within the
intitialization path of `sessionmgr`. This is the only thing I could
find after looking at the code that would result in
AgentRunner having been destroyed prior to the async initialization task
running.

Bug: 115009
Change-Id: I050afc1288f2f43e06bf7ede970ab94f5516c98c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770407
Commit-Queue: Gabe Schine <thatguy@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>