[ffx][fho] Macro hygiene improvements
There were a few places where fho macros expected things to be in the
available namespace. This makes it so they aren't anymore, using full
paths for all but one of them.

(The one unused one is `ffx_core::Injector` in the compatibility
macro, which is important because currently we auto-inject ffx_core into
all plugin deps and if this isn't used somewhere the tool will fail to
build for not using it -- this can be fixed later)

Change-Id: I20a982f27831098de2b4510fa0ab9d30c273b50a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771112
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>