[ermine] Use new AccountManager API for auth and enroll
This is a rough implementation of migrating to the new account manager
API behind a build flag. The goals is get consensus on the approach and
proper usage of the new account manager API. Once we have consensus, I
can fix the remaining fit, polish and test issues.

Bug: 113482
Change-Id: I1e95161bb6f2500e30b36b805b317c6b92efbaba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765624
Reviewed-by: Charles Whitten <cwhitten@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>