[fshost] fix nand device matching
When a NandDevice was processed, it caused an error in matching, because
the BootpartMatcher was tried first, which called get_block_info, which
returned an error for NandDevices.

To fix this, and to prevent similar logic bugs, break the matching
process into two steps - matching and processing. This applies some
invariants to the matching step - it's infallible, it can't mutate the
matcher, and it can't mutate the environment. Once something is matched,
we perform the fallible mutation, and cause matching to fail if we get
an error.

Test: fshost_rust_tests
Change-Id: I491ee0ce7c03d5ede46857f601f1e722699a8f2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771324
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>