[fs_management] one less error for detect_disk_format
Instead of returning an error when the header size we want to read is
larger than the block device we are checking, just say the content is
Unknown. This is closer to the cpp version (and arguably reality).

Change-Id: I2e6f5226596cdfd864468854d55da23edf8c5f2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771322
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>