[zither] Add support for vector syscall parameters
This change adds support for the use of `vector`s in the current FIDL
syscall model: such a parameter is decomposed as separate pointer and
length parameters.

To avoid having to generate C bindings for structs with vectors that are
not syscall request/response structs, we update the examples of this in
library zx, which Zither must be able to successfully interpret,
replacing these types with empty structs for now. Nothing currently
relies on these structs being properly interpreted at the moment, and
they are likely to be redesigned as this effort turns to the question of
how to model syscall data types in a first-class way.

Bug: 110294
Change-Id: Idc2b1946f756c2e5a1a992999c81fa926a8b924b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771363
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>