[memfs] Add client side zx::stream support
This change adds a new memfs build target that supports client side
streams. The implementation depends on zx_pager_query_vmo_stats which is
part of next_vdso and requires components that uses it to enable
use_next_vdso.

VnodeFile doesn't need to implement VmoRead because of the
supply_zero_offset in the kernel. All VnodeFile VMOs start with a size
of 0 and grow so supply_zero_offset will indicate that the zero page can
be supplied for all pages.

Bug: 95299
Change-Id: I07be91703adfe5648759cf152a2fc837fb050e40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764249
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Chris Tam <godtamit@google.com>