[build][bazel] Use @platforms//os:fuchsia directly.
Now that both the platform build and OOT SDK rules use the 0.0.6
release of the third-party Bazel platforms repository, which includes
a proper definition for Fuchsia, use it directly.

This is a followup to the sdk-integration CL at [1] which already
did this for the SDK rules. Note that this does not change the way
our platform definitions and toolchain resolution works.

[1] https://fuchsia-review.googlesource.com/c/sdk-integration/+/758284

Bug: None
Change-Id: I1f21d950cc49157eaca70b5fe60cc238265ce80c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770888
Commit-Queue: David Turner <digit@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>