[starnix] Update binder task API related to files
The previous API was exposing the FdTable, which is not relevant for a
RemoteBinder. Instead, expose the 3 file operation that are needed:
closing a file, exporting a file, importing a file.

Also update the behaviour when accessing files on a RemoteBinder and
return an error instead of doing a panic.

Change-Id: I4d89e10bc27e05180a61bab0d7fae380ef7c6ab1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770890
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>