[starnix] Fix lock ordering issue
If a matched value has a lock, the lock will be kept during the full
scope of the match. In that particular case, it generated a lock
inversion. This CL replaces the problematic match with a if/else.

Change-Id: I6b1c631678fb62ee439c358ae8f8aeb56a538cd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770891
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>