[codec] Remove video/h264-multi decoder
Currently both the video/h264 and video/h264-multi currently construct
the same underlying decoder implementation for both amlogic and x64. The
reason for this distinction was historical and no longer required. Tests
were renamed to remove the reference to the "multi" decoder. In some
cases, the same test was the same for the video/h264 and the
video/h264-multi decoder. In those cases the test was removed.

Change-Id: Icc32245de332be67c3d59e202236f60addbef424
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769622
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>