[kernel][vm] Move list node to DiscardableVmoTracker
Instead of tracking VmCowPages* in the discardable candidate lists,
track the DiscardableVmoTracker*. We will still be able to upgrade to a
RefPtr<VmCowPages> from the list node since DiscardableVmoTracker has a
raw back pointer to the owning VmCowPages.

This CL concludes the refactor to split up discardable metadata out of
VmCowPages. The size of the VmCowPagesContainer drops from 352 to 320
bytes as a result, and the DiscardableVmoTracker is 48 bytes. In other
words, the refactor saves 32 bytes of kernel metadata per
non-discardable VMO. The majority of VMOs used in practice are
non-discardable.

Run-All-Tests: true
Change-Id: I8dd0f50fdfdc98e29e11a91cebbb6e240ab23ef1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765045
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>