[fshost] support empty zxcrypt in integration tests
In the integration test fixture, support formatting the data partition
as zxcrypt with no filesystem inside. This changes adding zxcrypt to be
more additive - instead of an option to remove zxcrypt, we add a builder
method to add it in. Because most tests want to make that decision when
adding a data filesystem format, add an argument to that builder method
for adding zxcrypt.

Also add a fixture method for testing for the test file on the data
partition, to clean up a common flow in tests.

Bug: 87511
Test: fx test fshost_integration_tests
Change-Id: I6b43b8adba4c3bb8581bda2369e4ca5fe9419a28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771325
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>