[microbenchmarks] Generalise plumbing for same/different-CPU variants of tests
In an earlier change, plumbing was added so that the "ChannelPort" IPC
round-trip test could be run with same-CPU and different-CPU variants
(i.e. with threads pinned to the same or different CPUs).

Generalise the way that the parameters are passed so that this can
work for any of the test cases that use the ThreadOrProcess helper
class.

However, we don't enable that for the other test cases for now, to
avoid creating excessive numbers of test variants.  This plumbing just
allows interested people to see the results for other test variants by
adding additional calls to RegisterTestMultiProcSameDiffCpu().

Bug: 45802
Change-Id: Ib14ebb4d199fd1acdea724a922f901d793557273
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765048
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>