[boringssl] Disable ssl_test until upstream fixes expired cert
A particular certificate inlined into the test just expired, and this is
causing a test from the boringssl suite to fail consistently.  Disable
the test until upstream can make it pass again.

Ideally, we'd disable just the one test, but due to how we integrate
//third_party/boringssl/src directly from the upstream host, it's
complicated for us to patch it directly.

Change-Id: If198879a3bce96f25f666bf5b0262326e082d61e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771905
Reviewed-by: Ali Zhang <alizhang@google.com>