[edid] Move edid to //src/graphics/display/lib; cleanup deps
This moves edid out from //zircon/system/ulib so that it works
better as a display library. The library should only be used
with display core driver and the header definitions may be used
on audio driver; we removed all the other drivers that refer to
this as a dep but don't use it.

Change-Id: I9cedb7400a6e8c3822ea0fdd1ceaea9fd634afa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771487
Reviewed-by: Victor Costan <costan@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>