[memory] Add the list of VMOs per category.
This is useful to understant how VMO are grouped.

Also change the test to make sure no entry is missing
in `processes`. Provious code would pass with an empty
list.

With this change the --debug_json becomes only useful
for debug as all the information appears in the machine
output.

Bug: 116229
Change-Id: I5334b6e3d056f6c83c1b9de8669e7bbd453aa347
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/772682
Reviewed-by: Jean-François Geyelin <jif@google.com>
Commit-Queue: Pierre Labatut <plabatut@google.com>