[emu] Add a --stage flag to `ffx emu`
This CL adds a flag which delineates the `ffx emu start` subcommand at
the end of the Staging phase. Like --dry-run for Configuration, this
flag instructs `ffx emu` to short-circuit the start subcommand and print
the intermediate results of the staging. All file operations are
completed as part of the staging step, such as embedding SSH keys in the
ZBI and resizing the FVM as they are copied into the emulation instance
directory.

Bug: 103121
Change-Id: Ifbc70933ef8a938856eda090ebd10465af5ba679
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764826
Reviewed-by: Matt Hamrick <mjhamrick@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>