[build][bazel] Fix platform_aibs() implementation function for Bazel6
Bazel 6 now requires that the `output` and `target_file` arguments of
a `repository_ctx.symlink()` be either both files or both directories,
otherwise an error will be produced.

This CL updates the implementation function of hte platform_aibs() rule
to deal with this properly.

Bug: 112021
Change-Id: Iec8a99d7e71fff568224caa2bb228d59081a3264
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/772744
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>