[emu] Ensure --commit bypasses PBM processing
This CL adjusts the code in pbm::make_configs() to explicitly bypass the
product bundle processing during configuration when a file is provided
by the user with the --config flag. Processing a product bundle in this
scenario is wasteful and inaccurate, since none of the data in that
product bundle are going to be used.

To address the discrepancy between the passed-in configuration and the
contents of the EmulatorConfiguration, we also add a value to the
EmulatorConfiguration to indicate when the --config flag was used. When
calling `ffx emu show`, warnings are printed to explain the potential
discrepancy to the user, and some outputs are refused altogether.

Bug: 100653
Change-Id: I888e2c79b5249c915914673d93139f502f568019
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764827
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Matt Hamrick <mjhamrick@google.com>