[emu] Fix --reuse to reload previous config
Previously, `ffx emu start --reuse` would only reuse the disk images
from the previous instance of the emulator. In this CL, we modify the
start-up code to read the previous instances serialized engine from disk
and reuse the configuration as well. This means the user does not need
to reapply all of the command line arguments, and the launcher does not
need to re-process the product bundle metadata anymore.

Bug: 100653
Change-Id: Ibaf25dacf8ab18983e9f65af979dba8679a4ef31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764828
Reviewed-by: Matt Hamrick <mjhamrick@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>