[docs] Update third party management doc
Now that we have tools such as auto_owners that update third party
metadata and rely on a standardized third party directory structure,
rephrase the documentation to emphasize that the //third_party/ code
layout is expected and not just recommended, and that the OWNERS and
README.fuchsia files are required.

Add a 'Before you start' section highlighting the OSRB process instead of listing it in 'Additional readings'.

Add more details about third party OWNERS and their responsibilities.

Move section on migrating third party code to current layout to the end of the documentation.

Change-Id: I0c0ab0c12f3bd94559933bf139165b7248d966f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771062
Reviewed-by: Chris Palmer <palmer@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>