[emu] Update emulator state on deserialize
This CL ensures the most recent status is displayed for an emulator
instance. Specifically, if an emulator was most recently in the
"Running" state, but the process has since terminated, this will update
the serialized state file to indicate the emulator is now in the
"Staged" state. This reuses the existing "is_running()" code to verify
the process is still active, and only updates and writes the serialized
file on termination.

Bug: 112577
Change-Id: I202739d2c34595800d6606f2749ebb5f2ab0c336
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768142
Reviewed-by: Matt Hamrick <mjhamrick@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>