[libc] Test globals for dynamically loaded libs are registered correctly
The dynamic loader is required to call __hwasan_library_loaded from the
hwasan runtime whenever a dso is loaded in order to register global tags
into shadow memory. This is done explicitly in the loader (via
__sanitizer_library_loaded) rather than
via a function call that accepts a list of globals that is called in
.init_array to resolve an order of initialization problem. With
constructors we can encounter the following problematic scenario:

 1) library A depends on library B and also interposes one of B's symbols
 2) B's constructors are called before A's (as required for correctness)
 3) during construction, B accesses one of its "own" globals (actually
    interposed by A) and triggers a HWASAN failure due to the initialization
    for A not having happened yet

This adds tests for asserting various situations in which globals can be
accessed via module constructors work correctly without crashing with
hwasan.

Bug: 57112
Change-Id: I0a20c10de72ba4f6b6fb89a75d70508fd79b64c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724143
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>