[zither][zx] Add support for @wrapped_return
@wrapped_return annotates a singleton, response body of a protocol
method used to define a syscall and indicates that the return type of
the syscall is actually the type of the wrapped parameter. The reason
for this workaround is that protocol methods must return a struct.

This sidesteps the need for any of heuristics employed by kazoo today.

Bug: 110294
Change-Id: I0d091d1fc7c124e6d604efbbbc5fae07302fe913
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771364
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>