[ffxutil] Wait a little after ffx stop to allow shutdown to complete.
Sometimes the ffx.log is still being written to even after the daemon is
stopped. Since we upload this with the CAS outputs on failures, this can
cause the bot to die because the file is being written to while it's
being uploaded. See
https://chromium-swarm.appspot.com/task?id=5f01becebb5f8310. We used to
try to pgrep for the ffx daemon to check if it's still running, but
pgrep is not available on the host anymore.

Bug: 92296
Change-Id: I25498f412f3e8ab271559f79e79c065409d84ab1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771904
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>