[superproject] Roll third_party/pigweed pw_rpc: Test a call with a callback going out of scope
Callbacks are stored in the call object. If the call object goes out of
scope while the callback is executing, there could be issues with
dangling references in the callback. This test is disabled for now since
it does NOT yet pass, but will once b/234876851 is resolved.

Original-Bug: b/234876851
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/102602
Original-Revision: 6675519b44c1f7dbe18aede87d07256cfe88c714
GitOrigin-RevId: 2df395c9662d76b0b34b143858d4d6c71e46dd26
Roller-URL: https://ci.chromium.org/b/8795373844405039841
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I4d2fdc097810305e30d187f01ea6bee6e3115efb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773504