[zither][kernel] Add (syscalls|kernel).inc entry support
This change introduces support for the macro-friendly "C decl"s of the
syscalls that we find in the kazoo-generated syscalls.inc and kernel.inc
(as well as all of the variations on cdecls.inc).

With this change, Zither's internal conception/encoding of the syscall
interface should be complete.

Bug: 110294
Change-Id: I0ab20bb43a6b65ebdc2c7959a1e0e2b6b4619ef2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771366
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>