[emu] Fix scoping issue between stages in emu
The recent changes to introduce a firm state machine in `ffx emu`
included additional output of interim command lines for debugging. WE
missed a scoping issue, where one of the interim values was printed and
then discarded instead of continuing on to the next stage, which was
resulting in some of the configuration getting lost before final launch.

This CL fixes the scoping issue, such that the staged files and port
maps are used by the final launch.

Bug: 116956
Change-Id: I52fd7cc185af444025768f1f6cc2c773d7ed0b9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773546
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Matt Hamrick <mjhamrick@google.com>