[gazelle] Launch privileged apps in collections.
This CL creates dedicated collections for the terminal and
shell_settings applications in the gazelle shell. When they are
proposed as elements, their components get created in these
collections.

This is necessary for least privilege - just because the terminal
should have access to something doesn't mean that _all_ applications
should.

We could have also done this with dynamic offers, but that would have
moved the capability lists into code/a new file format to worry about.

This also adds support for multiple collections to element_manager.

Bug: fxbug.dev/105828
Bug: fxbug.dev/113488

Change-Id: Ia2e805518755b879982fafb1a71b15b82ddd2ed2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771803
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Ben Wright <benwright@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>