[assembly] Infer blobfs compression from package mode
The blobfs is not compressed if we place the FVM in the ZBI, because the
ZBI compression will optimize the size instead. Instead of requiring
boards to provide this variable we will infer its value from the package
mode passed to image assembly.

Bug: 117111
Change-Id: Ice5d044a09587122392449d797c8a3851f446d25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773142
Reviewed-by: David Iserovich <diserovich@google.com>
Reviewed-by: Daniel Young <youngda@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>