[graphics][compute][spinel] Support .deviceID of 0
The Lavapipe driver is unique and has a .deviceID of 0 so update the
Spinel benchmark logic to default to the first physical device when
both the .vendorID and .deviceID are both 0.

Also a small edit to to core.h.

Tested benchmark locally with the Lavapipe driver.

Change-Id: I62e5de94650c1cd6a743672814fac81c918d1970
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773165
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>