[as370] Delete AS370 board file and boot shim
There is now a board file for Visalia, so delete AS370. This was
originally for the AS370 dev board, which is no longer used. The
package-image script for Visalia is copied to a new location and will be
picked up by a subsequent change to that board file.

Bug: 94099
Test: Flashed and paved Visalia, device-enumeration-test passed
Change-Id: I8f1637dcf01b5ad56654fbfd88aa89cd4fa586c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709024
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>