[bt][controllers] Create FIDL and Banjo Controllers
Create bt-host/controllers directory for implementations of
pw::bluetooth::Controller.

Create FidlController, an implementation of Controller that uses
the BtHci FIDL API.

Create BanjoController, an implementation of Controller that uses the
BtHci and BtVendor Banjo APIs.

FidlController and BanjoController are intended to replace HciWrapper
and DeviceWrapper in the future.

Bug: 112750
Test: bt-host-controllers-tests

Change-Id: I281f70486209d22a6db45917c2c3992870468e01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739651
Reviewed-by: Faraaz Sareshwala <fsareshwala@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>