[kernel][x86] refresh the intel microarchiture list with newer entries
Just an addition of a few more newer (and a couple we had missed before)
to the list of intel microarchitectures. Nothing really new, except
Raptor Lake and Tremont, which is pointed at the microarch just before
it. Saltwell is added for completion, but it's unlikely the kernel will
run properly on it since it's a x86-64-v1 core.

Change-Id: I685b456d1689036e330618a351d660ee455bfea0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762296
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>