[fidl][docs] Update Wire* type docs
Update the docs for the new C++ bindings to indicate that
fidl::WireRequest and fidl::WireEvent are type aliases, and refer to
fidl::WireResult instead of fidl::WireResponse (which is mostly not
user-visible). Provide a clearer description of what is available in
fidl::WireResult.

Change-Id: Ie95f46b67090d7d3e92712e0903035b3f6197872
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773625
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>