[base_resolver] Support anonymous subpackages
Anonymous base subpackages are added to the on device base packages
manifest (the data/static_packages file of the system_image package)
with name/variant "_do_not_use_${HASH}/0". From the perspective of the
device, these anonymous packages are now base packages and so will
appear in pkgfs/packages and pkgfs/versions. This allows the
base_resolver to resolve anonymous subpackages because now their hashes
will appear in the BasePackageIndex when it does the reverse lookup and
their package directories can be opened from pkgfs/packages. As base
packages, they will be protected from GC.

The pkgfs/packages/<dir> subdirectory of packages is now a reserved
package name. None of the package resolvers will resolve a package with
this name. (The name itself is subject to change, but is currently
chosen to clearly convey it should not be used, since this directlry is
likely to go away very soon.)

Bug: fxbug.dev/101492
Fixed: fxbug.dev/101213

Change-Id: Ibd3fa151af02e2466e505a48cdf64fba2fa28019
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769753
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>