[fs_test] Remove mmap append-only test
Mmap'ing a file that was opened in append mode is allowed by POSIX.

ext2, ext3, and ext4 support an append-only attribute which only allows
the file to opened in append mode when writing. Files with this
attribute can not be mmap'ed with PROT_WRITE. Fuchsia does not support
this attribute and this attribute was not being used in the test.

Bug: 63987
Change-Id: I826dae67924d8f6acfe6e5ef2317a54339c0d4ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773666
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Chris Suter <csuter@google.com>