[archivist] Fix bug in event stream v2 handling which would lead to unhandled events
During shutdown, we wouldn't drain the merged stream that was created in
EventSource. By creating one EventSource per event stream we ingest, we
take advantage of the logic in place in EventRouter that ensures that we
drain all the events already received when shutdown is requested.

Even with that fix (in PS3), I think there's still potential for flaking
due to a race while pushing events through `dispatcher.emit().await` and
the receiver being terminated. To remove this race, I made the channel
unbounded so that the code that  places the events into the internal buffer is
entirely syncrhonous removing the potential of the receiver going away
while we are placing events in the channel. I don't think making this an
unbounded channel will be a problem in practice since it'll be storing
handles and identities that we'll be storing when the event is processed.

Change-Id: I6452861806774fa4e6efe850800380186c93eef5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775103
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>