[build][sdk] Add missing variants to ElementTypes
This adds the following variants to the ElementTypes enum that are used
in the sdk but weren't present before. This was causing failures in ffx
trying to read the released manifest after switching to using this
library to parse things.

It also adds a catch-all Unknown variant that will avoid erroring on
future additions.

Change-Id: Id6687881daab927bd60865fef886fdb5e3c461fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775243
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>