[input] Use physical pixels for injector viewport
Prior to this change, we reported mouse events in physical pixels, but
defined the injector viewport in logical pixels. As a result, clients
saw mouse events that appeared to be "pre-scaled" after applying the
viewport-to-view transform. As of this change, scene manager defines the
injector viewport's coordinate space in physical pixels, and accounts
for the device pixel ratio in the viewport-to-view transform.

This change also makes gazelle's pointer event usage internally
consistent, using logical coordinates.

Test: mouse-input-test, touch-input-test, verified on device
Bug: 113471
Bug: 116627
Change-Id: I9ba9bd4081179068228a8b70ecb8cd113c9987ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771908
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Alexander Brusher <abrusher@google.com>