[kernel][vm] Batch page requests to page sources
When performing direct operations on a VMO, the kernel knows exactly the
range in which pages are needed. For pages with page sources, this
change enables sending page requests in batches instead of separately
page-by-page. Consequently, when performing an operation on N
non-present pages of a pager-backed VMO, this reduces the number of the
outbound page requests to the userpager by (N - 1), which significantly
improves performance on large range operations.

Run-All-Tests: true
Test: core-pager-writeback
Test: core-pager
Test: core-stream

Fixed: 109983
Bug: 102657
Bug: 82818

Change-Id: I660486e345bfb2aecac9dc1fe0866183069e8201
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708011
Commit-Queue: Chris Tam <godtamit@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>