[dfv2][compat] Unbind removes children
Have the UnbindOp function fully remove each of its children before it
unbinds the current device.

This goes through the proper flow of shutting down the NodeController,
Unbinding the child, and calling Release on the child.

This makes UnbindOp a function that returns a promise, instead of a
function that accepts a callback.

Bug: 117387
Change-Id: Iebc7a53bb8e8d8c443ceb13359457d7ccc242bc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776164
Reviewed-by: Ian Fisher <iafisher@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>