[ffx][build] Flag to build dualmode plugins as tools
Adds a config flag that can be used to make any plugins that have been
converted over into being dual-mode (either plugin or separate tool)
build as separate tools.

This isn't really workable for building distribution sdks yet, until
the work is done on ffx, so the config will be disabled for now but
an be used to make local development less painful and begin the
process of migrating things to separate subtools.

Fixed: 117304
Change-Id: I17bf6dbb9a5f3e7a69343b26a8277756e9852f59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775662
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>