[input] Log keymap changes
We are getting a steady trickle of reports of spurious
keymap changes.

It is self-evident that this shouldn't be happening.

Since we don't have a good idea as to how that comes
about, I'm upgrading the debug log line for keymap ID
updates, so we can at least understand when this is
likely to happen.

This log like should be printed only rarely: when
the input pipeline first asks for the keymap, and then
on updates only. Since there is a singleton input
pipeline this should mean that the log lines will
be rare.

Bug: 116681
Change-Id: If92633dab9a5cc2c35ce89c425bdd730fe30ef02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776930
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>