[ffx] All top level suite-only commands can be subtools
This makes the remaining plugins that are currently only acting as
gateways to further levels of commands buildable as separate subtools,
and enables their building when the correct flag is set.

This is possible without any real changes to their code because none of
them have any library code at the root and the adapters we have are able
to handle them without changes (unlike 'normal' plugins).

Bug: 117304
Change-Id: Ic9fe7db916eceb3e741f24b72760f62661042878
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776502
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>