[fshost] fix write_data_file on fxfs with legacy crypto
The write data file implementation in the rust fshost didn't work for
fxfs when it was configured for legacy crypto, because it would try to
open a volume with the wrong name.

To fix this, write_data_file is refactored to use serve_data from the
FshostEnvironment, instead of implementing it's own serving and
formatting logic.

This case wasn't caught by the integration tests. Extend them, so they
run the whole suite of tests against fxfs with legacy crypto.

Bug: 87511
Fixed: 116842
Test: fx test fshost_integration_tests
Change-Id: I770d8fa11cea247458f7d4d977f2aa2f9711fdf1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773686
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>