[ffx][product] use a single client for downloading
The client factory and its clients cache the access token. When
a new factory + client are made the access token is unknown, which
causes a request for a new token. In some environments the new
access tokens are rate limited. This Cl uses a single client for
all the actions in a single ffx invocation which not only helps
in those environments, but accelerates the download for all users.

Bug: 103331
Change-Id: I78071befab2c93c6e4e50719b5e4a8d96aeebbf3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775986
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Michael Brase <mbrase@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>