[ffx][repo] Start should err if server is running on unexpected port
This changes `ffx repository server start` to error out if after we
started it, the server is running on a port we didn't expect. This can
happen if someone changes the listen port in the config without
restarting the server.

Fixed: 117737
Change-Id: Ic71a13de8a6c0ccf4390b643dce9ada6f6d80bf1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778769
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>