[audio] Adjust Pinecrest output driver valid_bits
AudioCore is currently limited to the 24-in-32 sample
format, if the container is int32. This change adjusts
the Pinecrest driver from 32-in-32 to 24-in-32. This
doesn't change anything about the signal, it just
communicates a format that AudioCore will accept.

The in-progress refactor of AudioCore will address this
AudioCore limitation, and fxbug.dev/117743 tracks the
restoring of this driver to full AUDIO_SAMPLE_FORMAT_32BIT
once that V2 audio refactor is landed.

Test: signal_generator on Pinecrest
Bug:  110904
Bug:  117743

Change-Id: Icda9a6aabc120374db8d242c4695f6f23ce1967c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778306
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>