[fidl][fidlc] Track leading newlines for tokens
The fact that the |Token| class tracks the |previous_end()| is a bit of
an anti-pattern in the modern compiler. It is a vestige of days prior
the "gaps" between tokens tracked in the raw AST were filled with lots
of useful information - discarded tokens like colons, comments, and so
on.

Today, |previous_end()| contains only whitespace. Of whitespace, the
only kind we care about are newlines, which both the formatter and the
transformer would like to retain as the user wrote them. Other kinds of
newlines (tabs and spaces) are never useful to us, as the pretty-printer
always has an opinionated way to render them, regardless of what prior
input says.

This all points to removing |previous_end()| and tracking
|leading_newlines()| as a number instead. One side benefit of this is
that we can now retain newlines when performing transforms (this is
relevant to the protocol modifier transform, for example). Another is
that we can delete a lot of existing formatter code that deduced newline
counts by manually walking over the |previous_end()| when needed.

Change I0b0bac39fe14d4b81f8aa7190c9583ff1f75a0cd that follows this one
runs the formatter against the entire *.fidl corpus, ensuring that
no weird edge cases slipped through.

Test: fx test fidlc-test; global fidl-format run
Bug: 114357, 117642
Change-Id: I0cd8b12709809381809d94224670b992c675ee28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777926
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>