[netstack3] Allow listen backlog <= 0
Implement the POSIX requirement that calling listen with a negative
backlog is the same as calling it with a backlog of 0, and implement the
allowed behavior of imposing a minimum backlog size. While imposing a
minimum, also impose a maximum of 4096 to match Linux behavior.

Bug: 117163
Change-Id: Ie52cfa460a1296e4f7387e9f557b3e137ea1460b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773145
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>