[build][bazel] Fix platform_aibs() for Bazel 6 (again)
This is a reland of [1] which was reverted because while it fixed
an issue for Bazel 6, it introduced another one for Bazel 5.x.

This version hopefully should work for any version, by adding a
new utility function that deals specifically with differences in
behavior between Bazel 5 and 6.

[1] https://fuchsia-review.git.corp.google.com/c/fuchsia/+/772744

Bug: 112021
Change-Id: I02a0464ff53fa02a33d7539ff0945d4f6f8b2354
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778822
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>