[fidl][fidlc] Remove outdated visitor_unittest
This test previously served two functions: ensure that all visitors were
called, and ensure that the file could be pretty printed using this
method. The first case is covered well by `span_tests.cc`, as well as
numerous other classes that inherit from the |TreeVisitor|, like the
linter and transformer.

The second case is still technically possible, but relies on code that
is being removed. Since the new canonical "pretty print stuff using the
tree visitor" tool is the formatter which builds on top of the tree
visitation functionality, this test becomes redundant. On top of that,
it is subtle and hard to maintain.

Test: fx test fidlc-test
Bug: 117642
Change-Id: Id9b7842b14f9b3e90b1a9167574bbf0abca6e1a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777928
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>