[wlan][policy] Separate original and tracked AP state.
This change introduces `ApState` in the Policy client state machine,
which contains BSS information from an initial connection and
independently tracks specific operational data that changes over time,
such as signal statistics and channel.

As part of this change, the `Sequestered` wrapper is used to make
certain operations on data fields more explicit and obvious in order to
avoid common mistakes with respect to design contracts. In particular,
the `BssDescription` associated with `ScannedCandidate` is considered a
"black box" and should generally not be inspected except when
constructing AP state.

Test: fx test wlancfg-tests
Bug: 102297
Change-Id: I3615c34af3c9d7ca87235c08e415187ed127be93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750413
Reviewed-by: Marc Khouri <mnck@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>
Reviewed-by: Hayden Nix <haydennix@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>