[starnix] Clarify UnixSocket.
I found the query_events() code to be very confusing. This adds a
comment explaining how the data is stored.

Moves the implementation of UnixSocketInner.query_events() inline to
the only caller, UnixSocket.query_events(). There are three query_events
functions involved (UnixSocket, UnixSocketInner, and MessageQueue) which
each return slightly different things. The "inner" version returned only
the part of the answer that could be computed without locking the peer.
I found this subset of behavior hard to reason about and hard to
describe in a comment, so moved it inline so there is one computation in
one place for the whole thing.

Change-Id: Ie0e7a53db9b31a76c04f2d761da1b5d846c65760
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778986
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>