[ethernet][e1000] Extend local rx buffer pool in e1000 driver
Packet drops happened because of the lack of rx buffer pool size in
e1000 driver. This change increases the buffer pool size from 32 to 128
to get a better error-tolerant rate.

Note: netdevice migration won't solve this automatically since the root
cause is that higher layer in the system is not consuming packets at a
stably rapid rate to support small buffer size like 32, so the packet
drop will happen anyway. After the migration, it will happen when there
is no QueueRxSpace to give driver and hardware enough rx spaces and
hardware chooses to drop the buffer.

Bug: 117189
Test: "fx pave" speed is stably ~90MB/s with the new buffer pool size.

Change-Id: Ic5765115c0786d55b8a1a2e9434ba699969f5e20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778781
Reviewed-by: Joon-Hyuk Han <hanjh@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Vinicius Felizardo <felizardo@google.com>