[starnix] Resolve waiter race on netlink socket.
NetlinkSocket.wait_async had a race condition because it checked the
current events, then optionally added a wait on those events. If another
event came along, in the middle, it could be missed.

This patch moves the entire function into a single lock (preventing
taking the lock twice in quick succession as before) and eliminates the
race.

Change-Id: I1dbd5ec1e9a6198e25b8395868da7689781ef224
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778861
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>