[forensics] Remove misleading upload log
When the network becomes reachable we should not unblock reports unless
the reporting policy is kUpload. Unblocking the reports was causing us
to misleadingly log 'Uploading x reports on network reachable' even if
the reporting policy resulted in us not uploading the reports.

This change is not unit testable, as all it results in is a change in
logging. However, I was able to manually verify through a temporary unit
test that the log statement is no longer triggered under kArchive.

Bug: 115163

TESTED=`fx test //src/developer/forensics`

Change-Id: I9341a9c96eb50e167dc365d820428128b2050421
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779145
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Thomas Gales <tgales@google.com>