[kernel][port] Improve scalability of ports and Semaphore
The purpose of this change is to reduce thread lock contention caused
by ports.

PortDispatcher uses a Semaphore to wake threads that are blocked in
zx_port_wait.  Every packet generated results in a call to
Semaphore::Post and every call to zx_port_wait results in a call to
Semaphore::Wait.  Prior to this change, every Wait and Post operation
on Semaphore acquired the thread lock regardless of whether there was
any need to block (in the case of Wait) or any waiters to wake (in the
cast Post).

This change improves port scalability by eliminating the need to need
to acquire the thread lock in the following scenarios:

* zx_port_wait when the packet queue contains one or more packets.
* Queuing a packet when the port has no waiters.

Bug: 111882
Run-All-Tests: true
Change-Id: I8939854ad22d40b529f79727c8768cb774b9b53e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730292
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: John Grossman <johngro@google.com>