[starnix] Add tests for the MAP_GROWSDOWN / PROT_GROWSDOWN feature
This adds a set of tests for the {MAP,PROT}_GROWSDOWN feature commonly
used to manage stacks. The tests cover growing a region and the
interaction with mappings in or near the guard region. This feature is
unusual in that it changes the effective protection of virtual addresses
that are not contained within any mapping the tests use a combination of
memory accesses and MAP_FIXED_NOREPLACE mappings to probe the effective
state of the address space. PROT_GROWSDOWN is also notable in that it
applies to regions outside the specified area and can modify the future
protections of unmapped regions.

Bug: 97514
Change-Id: I1884c79a049ce85d58c9c7b4b588402ddb209abf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778983
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>