[starnix][syncio] Pin zxio storage
It is possible for zxio objects to be self-referential; for example, see
the zxio_datagram_socket_t's RouteCache. This means that code that
manages that underlying storage must ensure it is not moved, as that
would invalidate existing references.

Document this invariant in zxio, and enforce it in syncio by pinning the
storage for the zxio object on the heap before calling into any
initialization logic. Expose a safe API wrapping this logic to clients
such as starnix.

Test: fx test starnix_gvisor_tests
Change-Id: Ib71fa78e32a8f7eb05c2777b6583a32249777ff6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778753
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>