[outgoing] Add |AddProtocol| unique_ptr overload
The |AddProtocol| overloads that take a raw pointer to
|fidl::Server<T>| or |fidl::WireServer<T>| are unsafe in
that they require users to coordinate the lifetime of the
servers with the |OutgoingDirectory| object. This new overload,
takes a |std::unique_ptr<ServerImpl>| which is safer because
the |OutgoingDirectory| owns and can ensure that the server is
destroyed and active connections are closed before going out of scope.

Change-Id: Ia1b798befe92d4ffb75784489971894cfca24081
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775483
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Gary Bressler <geb@google.com>