[xhci] Remove TRBPromise when TRB is unused
TRBPromise is often used throughout the driver for convenience, but the
TRB contained in the promise is unused and unnecessary. This makes it
harder to understand what the code is doing. In the interests of
optimising for the reader instead of the writer, this CL removes those
unnecessary uses of TRBPromise and replaces them with promise<void,
zx_status_t>.

Change-Id: I39c1fa2f7ff60a0a4d1af3b844d4b1fd837fa32c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765564
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>