[ffx] Make ffx target an explicit ABI revision from API level 10.
Targeting the latest ABI revision doesn't work well in practice as it
means that we will release a version of ffx that isn't compatible with
any previous version of the platform, and so users won't be able to use
this version of ffx to update their device. Targeting an historical ABI
isn't technically correct but it should provide a "grace period".

This is a temporary fix to unblock some users, a cleaner solution will
be provided shortly after this.

Bug: 117872
Change-Id: Iab1294d32931c05a7c21d2e19c7d57ef65f960ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779864
Reviewed-by: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: John Shamoon <johnshamoon@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Sébastien Marchand <sebmarchand@google.com>