[gazelle] Improve readability of shortcuts image.
The text on the post-it note isn't very legible
and there are some artifacts on the image. I made the
font larger and more bold, and got rid of the rotation
because it was creating some artifacts.

I'm assuming this is placeholder and so small tweaks
don't need to be run by UX.

Change-Id: Iec9c082e54a8f6c9dd6f2b9a30102ff8dc43cd02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780062
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Alexander Biggs <akbiggs@google.com>