[netstack3] Prevent overriding addr zone for TCP
Disallow setting the bound device for a socket if the socket already had
a device set and is bound or connected to an address that requires a
zone. This ensures that, once support for zoned addresses is added, any
zone set via binding or connecting can't be overridden by setting a
device on the socket.

Bug: 115524
Change-Id: I93824914c791c2ece7b0c683236884898bc972f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779944
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>