[network/tests] Explicitly depend on sync UDP
Introduce a Sync UDP variant of the netstack and depend on it explicitly
in the set of tests that need to run against both Fast UDP and Sync UDP.
This lets us flip the default in a child CL in an easily revertible way.

Once Fast UDP is turned on, we will remove the explicit Fast UDP
components, since they will then be redundant with the default.
Depending on whether we think we need the coverage, we might also remove
the Sync UDP components at some point.

Bug: 117436, 117883
Change-Id: If51c5ba4be648d17aba1ed57aab44308f15b3e8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779671
Reviewed-by: Nick Brown <nickbrow@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>