[doc-checker] Fixing references to topaz dart
This fixes links that point to the removed project "topaz" and
to in-tree Dart build rule files that have been moved.

The exact contents of the Dart build configuration is most likely
out of date, but is outside the scope of this change.

Bug: 117653
Change-Id: I3b2e9ddce5475770cdaa9cc5921ca032b1fe18b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780622
Reviewed-by: Kayce Basques <kayce@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Kayce Basques <kayce@google.com>