[doc-checker] Fixing  ambiguous links in md fragments
The presubmit program `doc-checker` checks links to local documentation
using filepaths. To enable this, links to local files should use file
paths. This is complicated by the use of included markdown fragments.
Normal practice is to place these fragments in a `_common` directory
and use `<< >>` to included them in a markdown file. When published,
this inclusion is done before links are rendered, so the links in the
included file are relative to the "host" file of the fragment. This
breaks the file path checking in `doc-checker` so full file paths are
preferred.

Bug: 117653
Change-Id: I2c86386550ff0b080591609d31d02e68c4f2a214
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780623
Reviewed-by: Kayce Basques <kayce@google.com>
Fuchsia-Auto-Submit: Kayce Basques <kayce@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>