[ffx][sdk] Add ffx_tool metadata parsing to ffx_sdk
In addition to just adding parsing of ffx_tool to the sdk, this also
restructures some things about how sdk parsing is done including
eliminating the really awkward function-trait based dependency injection
for testing things. In my opinion this simplifies a lot of the code
paths and makes things less weird.

That does mean it needs the sdk metadata to actually exist somewhere on
disk though, so this also moves the test data out into separate files
that are then referred to in tests.

This still needs to be wired up into fho for fxb/116732 to be complete.

Bug: 116732
Change-Id: I251cb750f770b6bc7221311633d437d1c758a3f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778777
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>