[ffx][fho] Validate fho metadata against a schema
Adds mechanisms for doing json schema validation of the fho metadata
without adding a big dependency on schemars. This is modelled after the
way the //build/sdk/meta libs do it, though it's a bit simpler.

The schema will be useful for writing relevant rfcs, and later we might
want to add the main ffx program validating the schema above and beyond
the serde-based validation already done.

Bug: 116730
Change-Id: Ieec12d312affce3ca45e2edcf8ca22d9eff03a92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779827
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>