[libpcap_test] Read instead of poll before pcap
The libpcap tests depend upon a packet being processed by the Netstack
before asserting on the results of `pcap_dispatch`. In preparation for
the launch of Fast UDP, which makes I/O asynchronous from the
perspective of the client, we landed https://fxrev.dev//713448 to `poll`
for POLLIN on the receiver before invoking `pcap_dispatch`. This breaks
when multiple packets are sent, because the first packet is never read
out from the receiver. To address this shortcoming, simply read out the
packet rather than polling.

Fixed: 117996
Multiply: libpcap-test
Change-Id: Id511acf4be9ea4f1d22c29b731276c4712bad296
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780145
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>