[events] Hooks::dispatch cannot fail
Currently, hooks are allowed to fail, which in turn can fail
the action that dispatched the event.

This should not be allowed. Hooks should not be able to affect
control flow. They are a unidirectional notification system
of important changes that occur to components.

If a hook produces an error, a warning message is printed out
by component manager, but control flow is not affected.

There are two tests that relied on a hook's ability to change
control flow: `destroy_error` and `shutdown_error`.

To work around this, we have introduced the ability to mock
results of actions in `ActionSet`. Instead of creating hooks
that fail an action, we now mock a failure for an action
instead.

Bug: 116855
Change-Id: I073b646e4d9049b7523c9844a8829cfb7c462349
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778752
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>