[assembly] Move component_manager_bootfs_config out of common
This config is not used in any real products, therefore it should not be
part of the common set of platform AIBs. Each product that wants it
should add it explicitly in their product gni definition until there is
a proper assembly feature to enable it.

Change-Id: I828f02fa520142110deab6e2bd9d0cfc847fb4d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780863
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>