[rust-3p] Move rustyline from mirrors/ to forks/
It isn't possible to soft-transition rustyline from nix 0.20 to 0.26.
It's //tp/rc/compat/rustyline/BUILD.gn is in-tree and the code is a
separate repo.

Rather than attempt a hard-transition, it seems better to move rustyline
in-tree and allow us to remove the separate mirror repo.

Once rolled, we should look at removing the rustyline petal from flower
and transition the old repo to read-only.

Bug: 117987

Change-Id: I4e6bf5210ece399ed06c4ea6fcbe1093e940444f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780202
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>