[zx-rs] {Channel,Event(Pair),Port,Timer}::create are infallible.
These functions no longer return a result, panicking if they
encounter an error.

The syscalls wrapped by these 5 constructors have in common that
they can only fail if either:

1. the kernel runs out of memory to create new objects
2. input or output pointers are invalid
3. non-zero options are provided
4. the process' job policy disallows it

(1) is explicitly mentioned in the syscall docs as a case that
userspace should not attempt to handle, and in the future the kernel
may not even return an error for it.

(2) and (3) are  prevented by construction in the Rust bindings to these
syscalls. The input/output pointers provided to the kernel are always
valid and the API of these functions do not allow passing non-zero
options to the syscalls.

(4) is possible, although in practice very few if any Rust programs run
on Fuchsia without job policy allowing creation of these 5 objects.
While one may want to be able to write such a program, I don't think that
is sufficient motivation to require the vast majority of Rust programs to
be prepared to handle the error.

Bug: 115383
Change-Id: I38cca2b70dd172a260dd414144728d0f4609f5fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759146
Commit-Queue: Adam Perry <adamperry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>