[wlan][policy] Fix flaky read of FIDL channel
Several of our tests put a message into one end of a FIDL channel,
then try to read from the other end with `stream.next()`. Occasionally,
the first attempt to run the `stream.next()` future returns
`Poll::Pending`, as though the kernel hasn't realized the message is
in the channel yet.

We chatted a bunch in the Rust chatroom, but couldn't find the _real_
underlying explanation. However, because the message is really in the
channel, retrying the attempt to read it works. There's a bunch of ways
to retry, but exec.run_singlethreaded works reliably in my local
testing. We could also loop (like we do elsewhere in the code), but
that's less elegant.

Fixed: 116134
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: I9485dddd0ff08be23497dce6522eb1673fc5c1fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781465
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>