[fidl][examples] Key value unknown interactions.
Add unknown interactions to the key-value-store examples. This changes
the main protocol to `open` and methods to `flexible`. This is mainly to
demonstrate that open/ajar exists rather than provide a specific example
of how to use them for protocol evolution; rather we just want them in
use since open/flexible is intended as the default in the future.

The add_iterator example does demonstrate a case where a protocol might
want to be declared as closed; the Iterator protocol is associated with
a single method which initiates iteration, so it is declared
closed/strict rather than open.

Bug: 88366
Change-Id: Ice63005d57f238f6214507dac2495ddc2552935f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780003
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>