[assembly] Add logic for WLAN assembly
Add logic to allow assembling different sets of WLAN packages based on
(1) level of privacy support and (2) whether the device is SoftMAC or
FullMAC. This logic is gated behind a flag, which is not set to true in
any platform configuration yet.

This additionally helps to unblock WLAN's USME migration project, as
WLAN needs to include different packages for FullMAC and SoftMAC during
the migration, and assembly has more meta information available to make
this decision.

Bug: 117249
Test: CQ
Change-Id: Id976fc67dc5bef02a1a3e75a39754030fbf990ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769011
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Kiet Tran <kiettran@google.com>