[audio] remove spurious command timeout
An unnecessary driver command timeout was being logged, when
moving the driver to the Start state. It does not seem that
this is the cause of the linked bug (since this timeout only
causes logging, not any change in behavior), but the log
message was found near the time of the problem.

BUG: b/262607079

Change-Id: I9d3e3c1d7e725a12762d3af1c0008d020c1f7e99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781662
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>