[fastboot] Migrate to |fidl::ServerBindingGroup|
This CL migrates |AddProtocol| usages where a raw pointer
to a FIDL server was passed as a handler. That overload is
deprecated. It is deprecated because the method overload can
hide subtle lifetime bugs. Instead, servers should use the
|unique_ptr| overload which gives ownership of the server to
|component::OutgoingDirectory|.

This was not done in this CL. The use cases migrated here passed
a `this` pointer. Thus it was hard to migrate to a |std::unique_ptr|
based approach without major refactoring. Consequently, this change opted
to instead leverage |fidl::ServerBindingGroup| to own connections, and
using the generated handles to handle connections.

Bug: 89036
Change-Id: Idc6d2228958d6801f70896e38ba78b18379fca79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780334
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>