[fidl][server_suite] Basic error reporting
Add a `WillTeardown` method to `Reporter`, such that some tests may
assert that the connection under test is torn down with the expected
reason. To test this works, we change the message_header_tests.cc to
assert the teardown reasons.

Dart doesn't have a way to observe errors on a binding. HLCPP doesn't
have the desired level of detail (it just reports a zx_status_t).
Changing the latter is much harder so I introduced a
`IsTeardownReasonSupported` queryable parameter for certain bindings
to opt-out of detailed teardown reason.

Finally, the definition of "unexpected message" doesn't quite match up
between the new C++ bindings and Rust/Dart. When parsing a FIDL
transaction header that has an invalid magic number, C++ treats that as
an unexpected message while the latter treats that as a decoding error.
While it would be good to standardize terminology, this CL doesn't
immediately do that and for now transforms the meaning of C++ unexpected
message to that of Rust/Dart inside cpp_utils.

Change-Id: I0afdc783c7f220c248e02d045aafe39a29d235b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780302
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Zachary Stewart <zstewart@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>