[devices] Migrate to |fidl::ServerBindingGroup|
This CL migrates |AddProtocol| usages where a raw pointer
to a FIDL server was passed as a handler. That overload is
deprecated. It is deprecated because the method overload can
hide subtle lifetime bugs. Instead, servers should use the
|unique_ptr| overload which gives ownership of the server to
|component::OutgoingDirectory|.

This was not done in this CL. The use cases migrated here passed
a `this` pointer. Thus it was hard to migrate to a |std::unique_ptr|
based approach without major refactoring. Consequently, this change opted
to instead leverage |fidl::ServerBindingGroup| to own connections, and
using the generated handles to handle connections.

Bug: 89036
Change-Id: I2597f129cb52876d747ee949cd6801f5591ee7a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780333
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>