[fshost] Support unformatted zxcrypt in netboot
When netboot is set, WriteDataFile only works if the data partition
happens to already be formatted as zxcrypt and if the driver is bound by
something else (i.e. the paver).

This means that SSH keys are not persisted when paving if the data image
payload is absent.  This is silly, since we don't actually care about
the data image payload's contents during paving (and prevents
fxrev.dev/781242 from landing).

Note that this also required a fix to the paver: The paver will bind
zxcrypt during partition initialization if needed (i.e. if the partition
in question is marked as a zxcrypt partition), but it never unbinds the
driver, which prevents fshost from functioning.  I could have fixed this
in fshost by attempting first to attach to the exisiting zxcrypt
instance, but that would be inherently racy, so I chose to fix it in the
paver instead.  There should not be any ill effects since the only
reason that the zxcrypt driver would need to be bound is for the SSH
keys to be written in, which is now handled by fshost via WriteDataFile.

Tested: integration, plus manually patching on top of the linked patch.
Change-Id: I51478e64e550597edc2b5cc466b20b7cf5177aab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781103
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>