[check-licenses] Insert proper license URL for Rust.
We want to link to the upstream copies of each license file.

The current strategy for doing this in check-licenses is somewhat rigid,
and it expects each project to only provide a single license file
or notice file. Rust provides two license files.

This CL adds a temporary bit of code to set the license URL for rust.
I'll remove this bit after designing a more flexible solution
that works for the entire repo.

Bug: 118257

Change-Id: I9404e3d72cd9cebff0eacbc6597c0ddff7d3fdb6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782272
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>