[check-licenses] Compress the large project.json test files.
project.json describes the workspace, and they tend to be
very very very large. One of the check-licenses tests attempts
to verify that we can accurately filter that file donw to only
the targets we care about. I committed this huge file to the
test directory, which is not ideal for several reasons
(more details in the bug).

This CL compresses the committed file, and decompresses it before
using it in the tests.

Fixed: 117994
Change-Id: I609f931f5ccb674a67aadc0685730eb6ed86f144
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783462
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>