[efi] Fix BootCpuid arithmetic error
The label `__start_BootCpuidLeaf` was defined for UEFI at an alignment
of 4 (the previous size of the contained data structure) rather than 8
(the current). As a result, we would sometime see an off-by-4 error in
the `arch::InitializeBootCpuid()` logic meant to populate this section.

This bug was consistently hit within the phys unit tests, which we
enable in this change.

Fixed: 115881
Change-Id: I7f17963f320f43ce4f53f35875a45286737fb0dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783582
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>