[fshost] Add a workaround for shredding the data volume
Shredding the data volume was broken on workstation whilst we're using
legacy crypto (non-hardware backed keys).  This adds a workaround: we
write a marker file that fshost will detect upon reboot and trigger a
reformat.  This change has only been made to the Rust version of fshost
which is the one used on workstation.  With this change, after erasing
all user data, the UI will no longer show the login screen after
rebooting.

During testing, I discovered that we would try and initialise the same
crypt service twice which would fail, so I've changed it so that the
crypt service is launched within a collection.  The Rust version should
destroy the components when they are no longer used, but the C++ version
will leave oprhaned components around, which is OK because we're hoping
to remove the C++ version soon.

Bug: 114384
Bug: 116488
Fixed: 117247
Change-Id: I28755be8db17a52d2013cc253d82bb5a09007914
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780266
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>