[tools][build] Refactor QEMU UEFI testing
This change updates how we perform boot tests through UEFI.

Before:
A UEFI filesystem/disk image was specified under the override name
"efi", and this was booted 'indirectly', akin to the UEFI Boot Manager
opting for the highest priority boot option on start-up. In this case,
a QEMU kernel was not passed.

After:
We support the previous mode of testing - but now with an image override
key of "efi_disk" - and add a more streamlined and ergonomic QEMU way of
booting a particular UEFI application directly: that is, by passing it
as a QEMU kernel (and so using the "qemu_kernel" override). So long as
the QEMU UEFI firmware is present, one can boot a UEFI application via
`-kernel`. This is more in line with how we boot through QEMU today and
has the advantages of being able to pass command line options (under
`-append`) as well as not needing to wrap the application in a FAT
filesystem image.

We keep the previous form of booting around as well, since we still need
to test applications booted from a more conventional UEFI filesystem
(and have one such test today). In this change, we move all instances of
`efi_test()` over to booting the new way.

Change-Id: I8a0dbd1ce1968ce9ee37149f87e7a1a43fa9a7f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774945
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>